backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0

# Data processing
pandas
sqlalchemy
psycopg2

# Embeddings
sentence-transformers>=2.2.2
SentencePiece


# LLM API
groq>=0.4.0
langchain_community
langchain 
langchain_groq

deep-translator
pycountry
langdetect 